Cambios y Prueba de la Conexion

+++ src/com/ucr/proyecto/domain/Empleado.java
@@ -32,7 +32,7 @@ public class Empleado implements Serializable {
     }
     
     // @param Parametro extra: Movimientos de cuenta
-    public Empleado(String usuario, String contrasena, int saldo, String numCuenta, String nombre,int codEmpleado, ArrayList movimientos) {
+    public Empleado(String usuario, String contrasena, double saldo, String numCuenta, String nombre,int codEmpleado, ArrayList movimientos) {
         this.usuario = usuario;
         this.contrasena = contrasena;
         this.saldo = saldo;
